% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{batchCount}
\alias{batchCount}
\title{Tests the number in each batch.}
\usage{
batchCount(dataS, batchSizeT)
}
\arguments{
\item{dataS}{Test dataset.}

\item{batchSizeT}{Batch size to test for.}
}
\value{
The ID of any batch that does not contain the specified number of samples.
}
\description{
Counts the number of samples that are in each of the batches.
}
\examples{
test <- batchCount(dataS=serumRand,batchSizeT=84)
}
