% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{emptyQC}
\alias{emptyQC}
\title{A dataset containing locations of empty QC samples.}
\format{
Empty QC samples
\describe{
\item{serumID}{Serum ID}
\item{rack}{Rack}
\item{row}{Row}
\item{col}{Column}
}
}
\description{
A dataset containing locations of empty QC samples.
}
\author{
Michelle Mellers
}
\keyword{data}
