% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{motherQC}
\alias{motherQC}
\title{A dataset containing locations of the mother QC samples.}
\format{
Mother QC samples
\describe{
\item{serumID}{Serum ID}
\item{rack}{Rack}
\item{row}{Row}
\item{col}{Column}
}
}
\description{
A dataset containing locations of the mother QC samples.
}
\author{
Michelle Mellers
}
\keyword{data}
