% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permRand-package.R
\docType{package}
\name{permRand-package}
\alias{permRand}
\alias{permRand-package}
\title{permRand: Permutation Randomization}
\description{
Provides randomization using permutation for applications. To provide a Quality Control (QC) check, QC samples can be randomized within strata. A second function allows for the ability to ”switch” samples to meet set requirements and perform a certain amount of minimization on these switches. The functions are flexible for users by specifying strata size and number of QC samples per strata. The randomization meets the following requirements • QC sample requirements: QC samples not adjacent, QC samples from same mother must follow certain patterns. • Matched sample sets must be within a single strata, and next to each other.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-mellers/permRand}
  \item Report bugs at \url{https://github.com/m-mellers/permRand/issues}
}

}
\author{
\strong{Maintainer}: Michelle Mellers \email{michelle.mellers.ctr@usuhs.edu} [copyright holder]

Authors:
\itemize{
  \item Thaddeus Haight \email{thaddeus.haight.ctr@usuhs.edu}
}

}
\keyword{internal}
