% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{QCMaster}
\alias{QCMaster}
\title{A dataset containing QC sample IDs and packing location.}
\format{
QC sample IDs and Packing Location
\describe{
\item{serumID}{Serum ID}
\item{rack}{Rack}
\item{row}{Row}
\item{col}{Column}
\item{event}{Sample Time}
\item{motherSerumID}{Mother Serum ID}
\item{motherRack}{Mother Rack}
\item{motherRow}{Mother Row}
\item{motherCol}{Mother Column}
\item{studyID}{Study ID}
}
}
\description{
A dataset containing QC sample IDs and packing location.
}
\author{
Michelle Mellers
}
\keyword{data}
