% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{serumIDs}
\alias{serumIDs}
\title{A dataset containing serum sample IDs linked with studyIDs.}
\format{
Serum Sample IDs linked with studyIDs
\describe{
\item{studyID}{Study Subject ID}
\item{event}{Sample Time}
\item{serumID}{Serum ID}
}
}
\description{
A dataset containing serum sample IDs linked with studyIDs.
}
\author{
Michelle Mellers
}
\keyword{data}
