% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{serumLoc}
\alias{serumLoc}
\title{A dataset containing the a packing list with serum samples.}
\format{
Packing list of serum samples
\describe{
\item{serumID}{Serum ID}
\item{rack}{Rack}
\item{row}{Row}
\item{col}{Column}
}
}
\description{
A dataset containing the a packing list with serum samples.
}
\author{
Michelle Mellers
}
\keyword{data}
