% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{serumMaster3}
\alias{serumMaster3}
\title{A dataset containing information for QC and serum samples.}
\format{
Serum and QC sample Information
\describe{
\item{serumID}{Serum ID}
\item{studyID}{Study ID}
\item{event}{Sample Time}
\item{rack}{Rack}
\item{row}{Row}
\item{col}{Column}
\item{ccID}{Group ID}
\item{caseControl}{Case or Control Indicator}
\item{QCsamp}{Indicates if QC sample}
}
}
\description{
A dataset containing information for QC and serum samples.
}
\author{
Michelle Mellers
}
\keyword{data}
