% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{serumSwitch}
\alias{serumSwitch}
\title{A dataset containing switched and randomized QC and serum samples.}
\format{
Switched and Randomized Serum and QC sample Information
\describe{
\item{serumID}{Serum ID}
\item{studyID}{Study ID}
\item{event}{Sample Time}
\item{ccID}{Group ID}
\item{caseControl}{Case or Control Indicator}
\item{QCsamp}{Indicates if QC sample}
\item{batchN}{Batch Number}
\item{loc}{Location}
\item{outoforder}{Marker if Switched}
}
}
\description{
A dataset containing switched and randomized QC and serum samples.
}
\author{
Michelle Mellers
}
\keyword{data}
