% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{serumSwitchP}
\alias{serumSwitchP}
\title{A dataset containing switched and randomized QC and serum samples
with packing locations.}
\format{
Switched and Randomized Serum and QC sample Information
\describe{
\item{serumID}{Serum ID}
\item{rack}{Rack}
\item{row}{Row}
\item{col}{Column}
\item{studyID}{Study ID}
\item{event}{Sample Time}
\item{ccID}{Group ID}
\item{caseControl}{Case or Control Indicator}
\item{QCsamp}{Indicates if QC sample}
\item{batchN}{Batch Number}
\item{loc}{Location}
\item{outoforder}{Marker if Switched}
}
}
\description{
A dataset containing switched and randomized QC and serum samples
with packing locations.
}
\author{
Michelle Mellers
}
\keyword{data}
