% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{testPair}
\alias{testPair}
\title{Ensures complete sets.}
\usage{
testPair(dataS)
}
\arguments{
\item{dataS}{Test dataset.}
}
\value{
The output reports any sets that are separated in the "loc".
}
\description{
Tests if sets are next to each other.  Any sets that are not next to each other are flagged.
}
\examples{
test <- testPair(dataS=serumRand)
}
