% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{testQCmatch}
\alias{testQCmatch}
\title{Tests QC matches}
\usage{
testQCmatch(dataS, numQCs, numMatch)
}
\arguments{
\item{dataS}{Randomized data.}

\item{numQCs}{Number of QCs specified per dataset.}

\item{numMatch}{Number of QC samples form a single mother within a batch.}
}
\value{
The output lists all batches with not enough QC sample sets or the QC samples do not come from the same mother.
}
\description{
Measures that every batch has at least the specified number of matching QC sample sets in a batch.
}
\examples{
test <- testQCmatch(dataS=serumRand,numQCs=4,numMatch=2)
}
