% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{testR}
\alias{testR}
\title{A dataset containing the example non-randomized serum sample file.}
\format{
A data frame with four datasets:
\describe{
\item{serumIDs}{Serum Sample IDs Linked with StudyIDs}
\item{serumLoc}{Serum ID packing list}
\item{emptyQC}{Empty QC containers packing list}
\item{motherQC}{QC samples to be aliquoted packing list}
}
}
\description{
A dataset containing the example non-randomized serum sample file.
}
\author{
Michelle Mellers
}
\keyword{data}
